rename: "interfaces" 패키지를 movie 패키지 하위로 옮김
