+++ src/main/java/nora/movlog/domain/movie/Movie.java
@@ -25,13 +25,13 @@ public class Movie {
     private String          runTime;   // 상영 시간
     private String          prdtYear;   // 제작 연도
 
-    @ManyToMany( cascade = CascadeType.ALL)
+    @ManyToMany(cascade = CascadeType.ALL)
     private Set<Nation>     nations;    // 제작 국가
-    @ManyToMany( cascade = CascadeType.ALL)
+    @ManyToMany(cascade = CascadeType.ALL)
     private Set<Genre>      genre;      // 장르
-    @ManyToMany ( cascade = CascadeType.ALL )
+    @ManyToMany (cascade = CascadeType.ALL)
     private Set<Director>   directors;  // 감독
-    @ManyToMany ( cascade = CascadeType.ALL )
+    @ManyToMany (cascade = CascadeType.ALL)
     private Set<Actor>      actors;     // 배우
 
     private WatchGrade      watchGrade; // 상영 등급
