+++ src/main/java/nora/movlog/repository/movie/interfaces/ActorRepository.java
@@ -1,4 +1,4 @@
-package nora.movlog.repository.interfaces;
+package nora.movlog.repository.movie.interfaces;
 
 import nora.movlog.domain.movie.Actor;
 import org.springframework.data.jpa.repository.JpaRepository;
