+++ src/main/java/nora/movlog/repository/movie/interfaces/GenreRepository.java
@@ -1,4 +1,4 @@
-package nora.movlog.repository.interfaces;
+package nora.movlog.repository.movie.interfaces;
 
 import nora.movlog.domain.movie.Genre;
 import org.springframework.data.jpa.repository.JpaRepository;
