+++ src/main/java/nora/movlog/service/movie/ActorService.java
@@ -2,7 +2,7 @@ package nora.movlog.service.movie;
 
 import lombok.RequiredArgsConstructor;
 import nora.movlog.domain.movie.Actor;
-import nora.movlog.repository.interfaces.ActorRepository;
+import nora.movlog.repository.movie.interfaces.ActorRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
