+++ src/main/java/nora/movlog/service/movie/DirectorService.java
@@ -2,7 +2,7 @@ package nora.movlog.service.movie;
 
 import lombok.RequiredArgsConstructor;
 import nora.movlog.domain.movie.Director;
-import nora.movlog.repository.interfaces.DirectorRepository;
+import nora.movlog.repository.movie.interfaces.DirectorRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
