+++ src/main/java/nora/movlog/service/movie/GenreService.java
@@ -2,7 +2,7 @@ package nora.movlog.service.movie;
 
 import lombok.RequiredArgsConstructor;
 import nora.movlog.domain.movie.Genre;
-import nora.movlog.repository.interfaces.GenreRepository;
+import nora.movlog.repository.movie.interfaces.GenreRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
