+++ src/main/java/nora/movlog/service/movie/NationService.java
@@ -2,7 +2,7 @@ package nora.movlog.service.movie;
 
 import lombok.RequiredArgsConstructor;
 import nora.movlog.domain.movie.Nation;
-import nora.movlog.repository.interfaces.NationRepository;
+import nora.movlog.repository.movie.interfaces.NationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
